return NAN if last packet was not received
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@11245 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/grb/conngrb.h
@@ -48,7 +48,7 @@ class ConnGrb:public rts2core::ConnNoSend
 		virtual void connectionError (int last_data_size);
 		virtual int receive (fd_set * set);
 
-		int lastPacket ();
+		double lastPacket ();
 		double delta ();
 		char *lastTarget ();
 		double lastTargetTime () { return last_target_time; }
